requirements.txt
Django~=5.1.5
pillow==11.1.0
django-tailwind==3.8.0
django-browser-reload==1.17.0
django-tinymce==4.1.0
